+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/psi/impl/xquery/XQueryCompElemConstructorPsiImpl.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 Reece H. Dunn
+ * Copyright (C) 2016, 2019 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,16 @@ package uk.co.reecedunn.intellij.plugin.xquery.psi.impl.xquery
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
+import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathEQName
+import uk.co.reecedunn.intellij.plugin.xpath.model.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryCompElemConstructor
+import uk.co.reecedunn.intellij.plugin.xquery.model.XQueryElement
 
-class XQueryCompElemConstructorPsiImpl(node: ASTNode) : ASTWrapperPsiElement(node), XQueryCompElemConstructor
+class XQueryCompElemConstructorPsiImpl(node: ASTNode) :
+    ASTWrapperPsiElement(node), XQueryCompElemConstructor, XQueryElement {
+    // region XQueryElement
+
+    override val nodeName get(): XsQNameValue? = findChildByClass(XPathEQName::class.java) as? XsQNameValue
+
+    // endregion
+}